ENC: `iohyve list` now displays the pool in which the guest belongs.
Results:
````
root@bhost:~ # iohyve list
Guest         VMM?  Running  rcboot?  Pool       Description
guestsystem   NO    NO       NO       apool      testing multi-disk
debian        YES   YES      NO       zpool      Ubiquiti mFi Controller Appliance
guestsystem1  NO    NO       NO       zpool      Tue Apr  5 16:02:36 MDT 2016
````

Basically duplicate names on different pools are bad, so don't do
drugs kids mmmkay. And here is why:

This commit will reveal a _insert-word_ with `iohyve`. If there are
two guests of the same name on different pools, both guests will display
`"YES"` for `VMM` and `Running` when only one is on. This is really
not a concern as fundentmentally `iohyve` would need to make many
changes for guests of the same name on different pools to work.
Currently `iohyve` does not include the pool name in the `VMM` instance,
nor does it `pgrep` for the pool when looking for a process. As a
matter of throughoutness, only the guest on the alphabetically first
pool is allowed to start. Two guests with the same name would only
occur from a `zfs send | recv` or an `iohyve rename` in which the
guest was on the alphabetically second (or more) pool (currently
there is not a check to see if the rename is itself (ie. `iohyve
rename bguest bguest`.) Quiet a few small changes can be done to
negate an accidential occurance of duplicate names with `iohyve`.

"The work continues." ~ Qyburn