Enh: Rewrite of info() to use flags for verbosity
Complete rewrite of iohyve info to use flags to
change the verbosity level of the output. Uses
flags -p -v -s -d and -l which must be used as one
series of characters after the "-". So "-dsl", "-v"
"-dvpls" all work, "-d -v -s -l -p" will not work.
Order of flags is not important.

 -p is for pool

 -v is for verbose which adds the fields: OS,
    Loader, Tap, and Con.

 -s is for status which add the fields: VMM,
    Running, and rcboot.

 -d is for description

 -l is for less and pipes the final output to less

The way this is written, extra characters are
ignored.