Enh: Reworked output for list() to use flags for verbosity
Reworked iohyve list to use flags. Removed description
from default output. With the use of flags the
description is diplayed. Flags -v -d and -l have been
added and are described below. The flags must be used
as a series of characters after the "-". So "-lvd",
"-v", "-dv" are correct syntax, however "-d -v -l"
is not. The order of the flags is not important.

-v is for verbose which adds the fields: Tap and Con

-d is for description

-l is for less and pipes the final output to less

The way this is written, extra characters are
ignored.