Fix issue# 210: iohyve create failing without pool name.
This contains my suggested fix for issue# 210, where iohyve create
failed the first time if no pool name was provided, and cpiso had been
used prior to create. Without this, all of the datasets for the isos get
listed in addition to the their parent dataset, and the pool name gets
listed multiple times instead of just once. With this, only the parent
dataset is counted. Future calls to create don't have the problem
regardless, because they look for the iohyve:name property rather than
looking for a dataset with /iohyve/ISO in its name.